+++ src/main/java/com/capacitybuilding/controllers/TrainingBean.java
@@ -53,7 +53,7 @@ public class TrainingBean implements TrainingBeanI {
         return trainings;
     }
 
-    public Training getTraining(int id) {
+    public Training getTraining(Long id) {
         Training training = entityManager.createQuery("FROM Training t WHERE t.id=:id", Training.class)
                 .setParameter("id", id)
                 .getResultList().get(0);
